pyproject.toml
[project]
name = "cricnws-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikhil Maheshwari"}
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "streamlit (>=1.50.0,<2.0.0)",
    "litellm (>=1.77.7,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "colorlog (>=6.9.0,<7.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastuuid (>=0.4.2,<1.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
